Add files via upload
🤴 KING SHIBA Introduction Pin 🐕
 
The King of the Shiba Kingdom has arrived to the BSC to provide his subjects protection for their investments and rewards for their fealty. Pledge yourself to the King and join an experienced team and doxxed dev in the next big memecoin. Sell all peasant shiba and floki tokens you hold and pay tribute to the King of Shibas!
 
🚨 Presale Information 🚨

50 BNB Private Sale that goes 100% to the marketing wallet (filled)

200 BNB Presale through Pinksale on Monday the 18th at 3PM UTC

The whitelist for the presale has already filled. Whitelisted wallets have the first opportunity to fill the presale with a max contribution of 0.5 BNB. If you are a Dr. Floki holder your wallet is automatically included in the whitelist. If the presale fails to fill from the whitelist (due to the small max contribution and limited whitelist), it will be open to the public.

🚀 Launch Information 🚀

We are LAUNCHING Tuesday the 19th at 5PM UTC. Launch is the best time to buy! The presale is small with a small max contribution, and there will be alot of resources in the marketing wallet to bring huge awareness to King Shiba. 

💰 Tokenomics 💰

2% Reflections
5% Marketing
3% Liquidity
2% Development for Staking System
5% Daily Burn until we reach 50% Burn!

 Buy/Sell Tax is 12%

⚡️ Features ⚡️

👉 EverOwn listing on launch day
👉 CMC and CG
👉 Influencer push on twitter and telegram call channels
👉 AMA’s in large crypto investing groups
👉 CEX Exchange listing 2 weeks from launch
👉 Staking KINGSHIBA dApp coming soon to Website
👉 Happy Hours with 0% trading tax
 
🔭 FIND OUT MORE 🔭

📖 Woofpaper: https://bit.ly/3FPSnW9

🌐 Website: https://www.kingshibaofficial.com

✉️ Telegram Link: https://t.me/KINGOFSHIBA

🐥 Twitter: https://twitter.com/KINGSHIBABSC